Add Prettier Config
This just runs on .mdx files for now. We can potentially expand to other files in future.

Prettier Config rules:
- https://prettier.io/docs/en/options.html#print-width
uses longer line lengths than the default (80). This means less changes in the codebase and makes things easier to read imo.
- https://prettier.io/docs/en/options.html#bracket-line
This is a good one which can make code sample appear nicer. see example in URL above to see.