#406: Part 1 - Restructure the modus web components repo.
This patch restructures the modus web components repo so that
there is a dedicated stencil-workspace for stencil development.
This is done so that we can clearly separate the stencil code
from the automatically generated react and angular components.

When React and Angular are implemented they will be automatically
generated and added to angular-workspace and react-workspace
respectively.

* Updated github actions to refer to the new stencil location
* Updated .prettierrc to force single quotes which is consistent
  with our lint settings
* Moved all stencil source code into ./stencil-workspace
* Moved storybook code along with stencil code to simplify the
  build script changes