Add GitHub's Super Linter Action
Add GitHub's Super Linter Action to run on each PR. This checks using linters for most file-types and is very helpful in catching issues with files we don't already lint.
REF: https://github.com/github/super-linter
We already use this on most other Modus projects and found it very helpful. We can always disable linting for a particular type if it becomes annoying. I usually disable Markdown and Editorconfig linting as they can be very strict and not too helpful.