Automate Release Notes Generation
This automates the Release Notes Generation. It is completely configurable and editable. You can manually change the text at any time. This will help avoid merge conflicts with CHANGELOG.md and save time.

When/if this is the merged the CHANGELOG.md can be updated with a link to check https://github.com/trimble-oss/modus-web-components/releases for latest changes