Pagination: Improved Button Sizes (#2047)
The next and previous buttons were too wide before, with this PR they are all square.
The horizontal padding for large pagination buttons has been decreased to make them square by default (for page 1), for larger numbers the buttons won't always be square. This change isn't perfect, but it's a big improvement.